+++ easypost/carrier_account.py
@@ -1,8 +1,10 @@
 from typing import (
+    Any,
     List,
     Optional,
 )
 
+from easypost.constant import _CARRIER_ACCOUNT_TYPES_WITH_CUSTOM_WORKFLOWS
 from easypost.easypost_object import convert_to_easypost_object
 from easypost.requestor import (
     RequestMethod,
@@ -16,6 +18,14 @@ from easypost.resource import (
 )
 
 
+def _select_carrier_account_creation_endpoint(carrier_account_type: Optional[Any]) -> str:
+    """Determines which API endpoint to use for the creation call."""
+    if carrier_account_type in _CARRIER_ACCOUNT_TYPES_WITH_CUSTOM_WORKFLOWS:
+        return "/carrier_accounts/register"
+
+    return "/carrier_accounts"
+
+
 class CarrierAccount(AllResource, CreateResource, UpdateResource, DeleteResource):
     @classmethod
     def types(cls, api_key: Optional[str] = None) -> List[str]:
@@ -24,11 +34,18 @@ class CarrierAccount(AllResource, CreateResource, UpdateResource, DeleteResource
         response, api_key = requestor.request(method=RequestMethod.GET, url="/carrier_types")
         return convert_to_easypost_object(response=response, api_key=api_key)
 
+    # Overrides CreateResource.create
     @classmethod
-    def register(cls, api_key: Optional[str] = None, **params) -> "CarrierAccount":
+    def create(cls, api_key: Optional[str] = None, **params) -> "CarrierAccount":
         """Creates a Carrier Account that requires custom registration (eg: FedEx & UPS)."""
         requestor = Requestor(local_api_key=api_key)
-        url = f"{cls.class_url()}/register"
+
+        carrier_account_type = params.get("type")
+
+        if carrier_account_type is None:
+            raise ValueError("Missing required parameter 'type'")
+
+        url = _select_carrier_account_creation_endpoint(carrier_account_type=carrier_account_type)
         wrapped_params = {cls.snakecase_name(): params}
         response, api_key = requestor.request(method=RequestMethod.POST, url=url, params=wrapped_params)
         return convert_to_easypost_object(response=response, api_key=api_key)