+++ easypost/constant.py
@@ -1,2 +1,7 @@
 SUPPORT_EMAIL = "support@easypost.com"
 TIMEOUT = 60
+
+_CARRIER_ACCOUNT_TYPES_WITH_CUSTOM_WORKFLOWS = [
+    "FedexAccount",
+    "UpsAccount",
+]